find_user/requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
psycopg==3.1.8
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
